feat: paymentMethod conversion
