--- src/components/Employee/PaymentMethod/BankAccountEdit.tsx
@@ -1,54 +1,0 @@
-import { useFormContext } from 'react-hook-form'
-import { useTranslation } from 'react-i18next'
-import type { BankAccountInputs } from './BankAccount'
-import { usePaymentMethod } from './usePaymentMethod'
-import { RadioGroupField, TextInputField } from '@/components/Common'
-import { PAYMENT_METHODS } from '@/shared/constants'
-
-export const BankAccountForm = () => {
-  const { mode, watchedType } = usePaymentMethod()
-  const { t } = useTranslation('Employee.PaymentMethod')
-  const { setValue } = useFormContext<BankAccountInputs>()
-
-  if ((mode !== 'ADD' && mode !== 'INITIAL') || watchedType === PAYMENT_METHODS.check) {
-    return
-  }
-  //Used by form schema to determine variant
-  setValue('hasBankPayload', true)
-
-  return (
-    <>
-      <TextInputField
-        name="name"
-        isRequired
-        label={t('nameLabel')}
-        errorMessage={t('validations.accountName')}
-      />
-
-      <TextInputField
-        name="routingNumber"
-        label={t('routingNumberLabel')}
-        isRequired
-        description={t('routingNumberDescription')}
-        errorMessage={t('validations.routingNumber')}
-      />
-
-      <TextInputField
-        name="accountNumber"
-        label={t('accountNumberLabel')}
-        errorMessage={t('validations.accountNumber')}
-        isRequired
-      />
-
-      <RadioGroupField
-        name="accountType"
-        isRequired
-        label={t('accountTypeLabel')}
-        options={[
-          { value: 'Checking', label: t('accountTypeChecking') },
-          { value: 'Savings', label: t('accountTypeSavings') },
-        ]}
-      />
-    </>
-  )
-}
