+++ src/components/Employee/PaymentMethod/onboarding/index.ts
@@ -1,0 +1,1 @@
+export { PaymentMethod, type PaymentMethodProps, PaymentMethodContextual } from './PaymentMethod'
